Refactored so both carousels use same ProductCard component
