+++ client/src/components/RelatedProducts/RelatedProducts.jsx
@@ -1,77 +1,8 @@
 import React, { useState, useEffect, useContext } from 'react';
 import { SharedContext } from '../../contexts/SharedContext';
-import styled from 'styled-components';
 import axios from 'axios';
-import RelatedProductCard from './RelatedProductCard';
 import Carousel from '../Carousel';
-
-const StyledRelatedProducts = styled.div`
-  border: 1px solid gray;
-  padding: 20px;
-  margin: 20px;
-  overflow: hidden;
-  width: 100%;
-  height: 100%;
-`
-
-const StyledCarouselContainer = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-`
-
-const StyledCarouselWrapper = styled.div`
-  display: flex;
-  width: 100%;
-  position: relative;
-`
-
-const StyledCarouselContentWrapper = styled.div`
-  overflow: hidden;
-  width: 100%;
-  height: 100%;
-`
-
-const StyledCarouselContent = styled.div`
-  display: flex;
-  transition: all 250ms linear;
-  -ms-overflow-style: none;
-  scrollbar-width: none;
-  width: calc(100% / 4);
-
-  &::-webkit-scrollbar {
-    display: none;
-  }
-
-  > * {
-    width: 100%;
-    flex-shrink: 0;
-    flex-grow: 1;
-  }
-`
-
-const Arrow = styled.div`
-  position: absolute;
-  z-index: 1;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 48px;
-  height: 48px;
-  border-radius: 24px;
-  background-color: white;
-  border: 1px solid #ddd;
-  text-align: center;
-  vertical-align: middle;
-  line-height: 48px;
-`
-
-const StyledLeftArrow = styled(Arrow)`
-  left: 0px;
-`
-
-const StyledRightArrow = styled(Arrow)`
-  right: 0px;
-`
+import ProductCard from '../ProductCard';
 
 const RelatedProducts = () => {
   const { productId } = useContext(SharedContext);
@@ -99,7 +30,7 @@ const RelatedProducts = () => {
       <h4>RELATED PRODUCTS</h4>
       {isLoaded ? <Carousel products={relatedProducts} mode='RelatedProducts'>
         {relatedProducts.map((product) => (
-          <RelatedProductCard key={product} product_id={product} />
+          <ProductCard key={product} product_id={product} mode='RelatedProducts'/>
         ))}
       </Carousel> : <div></div>}
     </>
